refactor(IT-Cotato#216): entity, repository 패키지 수정
